Rename mappings file
