+++ src/hapi/pipelines/utilities/mappings.py
